Allow preserving files on destroy
This change introduces the `preserve_on_destroy` option for the
local_file resource. This defaults to `false`, but when set to `true`
will ensure that the file is preserved when the resource is destroyed in
Terraform.

Resolves: #7

+++ local/resource_local_file.go
@@ -60,6 +60,13 @@ func resourceLocalFile() *schema.Resource {
 				Default:      "0777",
 				ValidateFunc: validateMode,
 			},
+			"preserve_on_destroy": {
+				Type:        schema.TypeBool,
+				Description: "If true, the file will be preserved on destroy.",
+				Optional:    true,
+				Default:     false,
+				ForceNew:    true,
+			},
 		},
 	}
 }
@@ -133,6 +140,8 @@ func resourceLocalFileCreate(d *schema.ResourceData, _ interface{}) error {
 }
 
 func resourceLocalFileDelete(d *schema.ResourceData, _ interface{}) error {
-	os.Remove(d.Get("filename").(string))
+	if !d.Get("preserve_on_destroy").(bool) {
+		os.Remove(d.Get("filename").(string))
+	}
 	return nil
 }
